requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
ultralytics==8.4.24
Pillow>=10.0.0
